pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-tts"
version = "0.2.0"
description = "Qwen3-TTS macOS Apple Silicon 対応・日本語ローカライズ版"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
license = { text = "Apache-2.0" }
authors = [{ name = "Alibaba Qwen Team" }]

dependencies = [
  "transformers==4.57.3",
  "accelerate==1.12.0",
  "gradio",
  "librosa",
  "torchaudio",
  "soundfile",
  "onnxruntime",
  "einops",
]

[project.optional-dependencies]
whisper = ["faster-whisper"]

[project.urls]
Homepage = "https://github.com/shiroemons/Qwen3-TTS-JP-mac"
Repository = "https://github.com/shiroemons/Qwen3-TTS-JP-mac"
Upstream = "https://github.com/QwenLM/Qwen3-TTS"

[project.scripts]
qwen-tts-demo = "qwen_tts.cli.demo:main"

[tool.setuptools]
packages = { find = { where = ["."] , include = ["qwen_tts*"] } }
include-package-data = true

[tool.setuptools.package-data]
qwen_tts = ["py.typed", "**/*.npz"]